(rigoberto)new:tasks.js(func para interactuar con el back)
